Closed #25; implementate api Manager loadSelectedFile

+++ backend/main.py
@@ -1,7 +1,6 @@
 import os
 import time
-from chatsql import ChatSQL
-from flask import Flask, jsonify
+import chatsql 
 from flask_cors import CORS
 from flask import Flask, request, jsonify
 from werkzeug.utils import secure_filename
@@ -17,8 +16,9 @@ app.config.from_object(__name__)
 # enable CORS
 CORS(app, resources={r'/*': {'origins': '*'}})
 
-current_file = None
-chatSql = None
+def allowed_file(filename):
+    return '.' in filename and \
+            filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
             
 @app.route('/upload', methods=['POST'])
 def upload_file():
@@ -44,7 +44,8 @@ def get_files():
         mtime = os.path.getmtime(path)
         date = time.ctime(mtime)
         name, extension = os.path.splitext(filename)
-        loaded = path == current_file
+        loaded = chatsql.verify_file_name(filename)
+        print(loaded)
         files.append({
             'name': name,
             'extension': extension,
@@ -56,18 +57,11 @@ def get_files():
 
 @app.route('/file/<filename>', methods=['GET'])
 def get_file_info(filename):
-    global current_file
-    global chatSql
-
     try:
-        current_file = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-        if os.path.exists(current_file):
-        
-            if (chatSql is not None):
-                del chatSql
-            
-            chatSql = ChatSQL(current_file)
-                    
+        file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
+        if os.path.exists(file_path):
+            chatsql.set_file_path(file_path)
+            chatsql.Init()
             return jsonify({'message': f'File {filename} found', 'found': True, 'filename': filename})
         else:
             return jsonify({'message': f'File {filename} does not exist', 'found': False})
@@ -77,7 +71,7 @@ def get_file_info(filename):
 @app.route('/getloadedfile', methods=['GET'])
 def getloadedfile():
     try:
-        return jsonify({'filepath': current_file}, success=True)
+        return jsonify({'filepath': chatsql.get_file_name()}, success=True)
     except Exception as e:
         return jsonify({'message': str(e), 'found': False})
     
@@ -85,12 +79,11 @@ def getloadedfile():
 def delete_file(filename):
     try:
         file_path = os.path.join(app.config['UPLOAD_FOLDER'], filename)
-        try:
+        if os.path.exists(file_path):
             os.remove(file_path)
-            return jsonify({'message': 'File Deleted'})
-        except OSError as er:
-            err_message = "Error: "+er.filename+" - "+er.strerror
-            return jsonify({'message': err_message})
+            return jsonify({'message': 'File successfully deleted'})
+        else:
+            return jsonify({'message': 'File does not exist'})
     except Exception as e:
         return jsonify({'message': str(e)})
